feat: entities services and repository

+++ src/main/java/fr/istic/m2/mcq_api/dto/AnswerDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class AnswerDTO {
-    private int questionId;
+    private Long questionId;
     private boolean isValid;
     private int nbrPoint;
 }
