+++ src/main/java/fr/istic/m2/mcq_api/dto/QuestionDTO.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class QuestionDTO {
-    private int qcmId;
-    private int levelId;
+    private Long qcmId;
+    private Long levelId;
     private boolean isActive;
     private int delay;
     private String title;
